+++ src/core/objective.jl
@@ -8,12 +8,12 @@
 """
     objective_mc_se
 """
-function objective_mc_se(pm::_PMs.AbstractPowerModel)
+function objective_mc_se(pm::_PMD.AbstractUnbalancedPowerModel)
     return JuMP.@objective(pm.model, Min,
     sum(
         sum(
-            sum(_PMs.var(pm, nw, :res, i)[idx] for idx in 1:length(_PMs.var(pm, nw, :res, i)) )
-        for i in _PMs.ids(pm, nw, :meas))
-    for (nw, nw_ref) in _PMs.nws(pm) )
+            sum(_PMD.var(pm, nw, :res, i)[idx] for idx in 1:length(_PMD.var(pm, nw, :res, i)) )
+        for i in _PMD.ids(pm, nw, :meas))
+    for (nw, nw_ref) in _PMD.nws(pm) )
     )
 end