+++ src/core/start_values_methods.jl
@@ -7,24 +7,18 @@
 ################################################################################
 """
     assign_start_to_variables!(data)
-
 This function gives the measurement value in the data dictionary
 as a starting value to its associated variable.
 """
 function assign_start_to_variables!(data::Dict{String, Any})
     for (_,meas) in data["meas"]
         if all(i -> i ∈ [_DST.Normal{Float64}, Float64], typeof.(meas["dst"]))
-            if string(meas["var"]) != "w"
-                data[string(meas["cmp"])][string(meas["cmp_id"])]["$(string(meas["var"]))_start"] = _DST.mean.(meas["dst"])
-            else
-                data[string(meas["cmp"])][string(meas["cmp_id"])]["$(string(meas["var"]))_start"] = _DST.mean.(meas["dst"])[1]
-            end
+            data[string(meas["cmp"])][string(meas["cmp_id"])]["$(string(meas["var"]))_start"] = _DST.mean.(meas["dst"])
         end
     end
 end
 """
     assign_start_to_variables!(data, start_values_source)
-
 This function assigns start values to the problem variables based on a dictionary where they are collected: start_values_source.
 This dictionary must have the form of a powerflow solution dictionary.
 """
@@ -40,7 +34,7 @@ function assign_start_to_variables!(data::Dict{String, Any}, start_values_source
                 data[msr_cmp][cmp_id]["$(msr_var)_start"] = start_values_source["solution"][msr_cmp][cmp_id][msr_var][1]
             end
         else
-            Memento.warn(_LOGGER, "$(msr_var) is not in $(start_values_source), possible formulation mismatch")
+            @warn "$(msr_var) is not in $(start_values_source), possible formulation mismatch"
         end
     end
 end