+++ src/core/utils.jl
@@ -11,10 +11,10 @@ Retrieves the id of component i. This is a tuple if the component is a branch. O
 """
 function get_cmp_id(pm, nw, i)
     if  _PMD.ref(pm, nw, :meas, i, "cmp") == :branch
-        branch_id = _PMs.ref(pm, nw, :meas, i, "cmp_id")
+        branch_id = _PMD.ref(pm, nw, :meas, i, "cmp_id")
         cmp_id = (branch_id, _PMD.ref(pm,nw,:branch, branch_id)["f_bus"], _PMD.ref(pm,nw,:branch,branch_id)["t_bus"])
     else
-        cmp_id = _PMs.ref(pm, nw, :meas, i, "cmp_id")
+        cmp_id = _PMD.ref(pm, nw, :meas, i, "cmp_id")
     end
     return cmp_id
 end
@@ -90,7 +90,7 @@ function assign_basic_individual_criteria!(data::Dict; chosen_criterion::String=
             data["crit"] = "mle"
         end
     else
-        Memento.error(_LOGGER, "Unrecognized data input.")
+        error("Unrecognized data input.")
     end
 end
 """
@@ -112,7 +112,7 @@ function assign_residual_ub!(data::Dict; chosen_upper_bound::Float64=100.0, resc
     elseif haskey(data, "dst")
         data["res_max"] = upp_bound
     else
-        Memento.error(_LOGGER, "Unrecognized data input.")
+        error("Unrecognized data input.")
     end
 end
 """
@@ -181,23 +181,32 @@ function reduce_single_phase_loadbuses!(data::Dict; exclude=[])
     for (bus_id,load_id,nr_conn_loads) in load_info
         load_connections = data["load"][load_id]["connections"]
         if nr_conn_loads == 1 && length(load_connections) == 1
-            perform_dimension_reduction(data, bus_id, load_connections)
+            perform_dimension_reduction!(data, bus_id, load_connections)
         elseif nr_conn_loads > 1 && bus_id ∉ already_checked
             push!(already_checked, bus_id)
             loads_at_bus_id = [load_id for (x, load_id, z) in load_info if x == bus_id]
             used_connections = unique([data["load"][lid]["connections"] for lid in loads_at_bus_id])
-            if length(used_connections) == 1 perform_dimension_reduction(data, bus_id, load_connections) end
+            if length(used_connections) == 1 perform_dimension_reduction!(data, bus_id, load_connections) end
         end
     end
 end
 "reduces the dimension of bus terminals and branches f_ and t_connections to match those of the connected load(s)"
-function perform_dimension_reduction(data::Dict, bus_id::Int64, load_connections)
+function perform_dimension_reduction!(data::Dict, bus_id::Int64, load_connections)
     data["bus"]["$bus_id"]["terminals"] = load_connections
     data["bus"]["$bus_id"]["grounded"] = data["bus"]["$bus_id"]["grounded"][load_connections]
     conn_branches = find_branch_t_bus(data["branch"], bus_id)
+
+    idx = load_connections[1]
+
     for br_id in conn_branches
-        data["branch"][br_id]["f_connections"] = load_connections
-        data["branch"][br_id]["t_connections"] = load_connections
+        branch = data["branch"][br_id]
+        for (key,value) in branch
+            if isa(value, Matrix)
+                branch[key] = reshape([value[idx,idx]], 1, 1)
+            elseif isa(value, Vector)
+                branch[key] = load_connections
+            end
+        end
     end
 end
 "returns a tuple with all loads' information. Every load is assigned a tuple with the following content: (bus the load is connected to, the load index, total number of loads connected to the same bus) "
@@ -219,13 +228,13 @@ function find_branch_t_bus(branches, bus_id)
     for (b, branch) in branches
         if branch["t_bus"] == bus_id push!(conn_branches, b) end
     end
-    !isempty(conn_branches) ? (return conn_branches) : Memento.error(_LOGGER, "Network graph is disconnected")
+    !isempty(conn_branches) ? (return conn_branches) : error("Network graph is disconnected")
 end
 """
-    function get_active_connections(pm::_PMs.AbstractPowerModel, nw::Int, cmp_type::Symbol, cmp_id::Int)
+    function get_active_connections(pm::_PMD.AbstractUnbalancedPowerModel, nw::Int, cmp_type::Symbol, cmp_id::Int)
 Returns the list of terminals, connections or t_ and f_connections, depending on the type of the component.
 """
-function get_active_connections(pm::_PMs.AbstractPowerModel, nw::Int, cmp_type::Symbol, cmp_id::Int)
+function get_active_connections(pm::_PMD.AbstractUnbalancedPowerModel, nw::Int, cmp_type::Symbol, cmp_id::Int)
     if cmp_type == :bus
        active_conn = _PMD.ref(pm, nw, :bus, cmp_id)["terminals"]
    elseif cmp_type ∈ [:gen, :load]
@@ -233,7 +242,7 @@ function get_active_connections(pm::_PMs.AbstractPowerModel, nw::Int, cmp_type::
    elseif cmp_type == :branch
        active_conn = intersect(_PMD.ref(pm, nw, :branch, cmp_id)["f_connections"], _PMD.ref(pm, nw, :branch, cmp_id)["t_connections"])
    else
-       Memento.error(_LOGGER, "Measurements for component of type $cmp_type are not supported")
+       error("Measurements for component of type $cmp_type are not supported")
    end
    return active_conn
 end