+++ src/form/reduced_ivr.jl
@@ -2,26 +2,26 @@
 #  Copyright 2020, Marta Vanin, Tom Van Acker                                  #
 ################################################################################
 # PowerModelsDistributionStateEstimation.jl                                    #
-# An extention package of PowerModels(Distribution).jl for Static Power System #
+# An extention package of PowerModelsDistribution.jl for Static Power System   #
 # State Estimation.                                                            #
 ################################################################################
-mutable struct ReducedIVRPowerModel <: _PMs.AbstractIVRModel _PMs.@pm_fields end
+mutable struct ReducedIVRUPowerModel <: _PMD.AbstractUnbalancedIVRModel _PMD.@pmd_fields end
 
 "only total current variables defined over the bus_arcs in PMD are considered: with no shunt admittance, these are
 equivalent to the series current defined over the branches."
-function variable_mc_branch_current(pm::ReducedIVRPowerModel; nw::Int=pm.cnw, bounded::Bool=true, report::Bool=true, kwargs...)
+function variable_mc_branch_current(pm::ReducedIVRUPowerModel; nw::Int=_IM.nw_id_default, bounded::Bool=true, report::Bool=true, kwargs...)
     _PMD.variable_mc_branch_current_real(pm, nw=nw, bounded=bounded, report=report; kwargs...)
     _PMD.variable_mc_branch_current_imaginary(pm, nw=nw, bounded=bounded, report=report; kwargs...)
 end
 
 "if the formulation is not reduced, it is delegated back to PMD"
-function variable_mc_branch_current(pm::_PMs.IVRPowerModel; nw::Int=pm.cnw, bounded::Bool=true, report::Bool=true, kwargs...)
+function variable_mc_branch_current(pm::_PMD.IVRUPowerModel; nw::Int=_IM.nw_id_default, bounded::Bool=true, report::Bool=true, kwargs...)
     _PMD.variable_mc_branch_current(pm, nw=nw, bounded=bounded, report=report; kwargs...)
 end
 
-"constraint_mc_generator_power is re-defined here because in PMD the same function only accepts pm::_PMs.IVRPowerModel.
+"constraint_mc_generator_power is re-defined here because in PMD the same function only accepts pm::_PMD.IVRUPowerModel.
 The content of the function is otherwise identical"
-function constraint_mc_generator_power(pm::ReducedIVRPowerModel, id::Int; nw::Int=pm.cnw, report::Bool=true, bounded::Bool=true)
+function constraint_mc_generator_power(pm::ReducedIVRUPowerModel, id::Int; nw::Int=_IM.nw_id_default, report::Bool=true, bounded::Bool=true)
 
     generator = _PMD.ref(pm, nw, :gen, id)
     bus = _PMD.ref(pm, nw,:bus, generator["gen_bus"])
@@ -39,9 +39,9 @@ function constraint_mc_generator_power(pm::ReducedIVRPowerModel, id::Int; nw::In
     end
 end
 
-"constraint_mc_generator_power_wye is re-defined here because in PMD the same function only accepts pm::_PMs.IVRPowerModel.
+"constraint_mc_generator_power_wye is re-defined here because in PMD the same function only accepts pm::_PMD.IVRUPowerModel.
 The content of the function is otherwise identical"
-function constraint_mc_generator_power_wye(pm::ReducedIVRPowerModel, nw::Int, id::Int, bus_id::Int, connections::Vector{Int}, pmin::Vector{<:Real}, pmax::Vector{<:Real}, qmin::Vector{<:Real}, qmax::Vector{<:Real}; report::Bool=true, bounded::Bool=true)
+function constraint_mc_generator_power_wye(pm::ReducedIVRUPowerModel, nw::Int, id::Int, bus_id::Int, connections::Vector{Int}, pmin::Vector{<:Real}, pmax::Vector{<:Real}, qmin::Vector{<:Real}, qmax::Vector{<:Real}; report::Bool=true, bounded::Bool=true)
 
     vr = _PMD.var(pm, nw, :vr, bus_id)
     vi = _PMD.var(pm, nw, :vi, bus_id)
@@ -77,9 +77,9 @@ function constraint_mc_generator_power_wye(pm::ReducedIVRPowerModel, nw::Int, id
    _PMD.var(pm, nw, :qg)[id] = JuMP.Containers.DenseAxisArray(qg, connections)
 
 end
-"constraint_mc_generator_power_delta is re-defined here because in PMD the same function only accepts pm::_PMs.IVRPowerModel.
+"constraint_mc_generator_power_delta is re-defined here because in PMD the same function only accepts pm::_PMD.IVRUPowerModel.
 The content of the function is otherwise identical"
-function constraint_mc_generator_power_delta(pm::ReducedIVRPowerModel, nw::Int, id::Int, bus_id::Int, connections::Vector{Int}, pmin::Vector{<:Real}, pmax::Vector{<:Real}, qmin::Vector{<:Real}, qmax::Vector{<:Real}; report::Bool=true, bounded::Bool=true)
+function constraint_mc_generator_power_delta(pm::ReducedIVRUPowerModel, nw::Int, id::Int, bus_id::Int, connections::Vector{Int}, pmin::Vector{<:Real}, pmax::Vector{<:Real}, qmin::Vector{<:Real}, qmax::Vector{<:Real}; report::Bool=true, bounded::Bool=true)
     vr = _PMD.var(pm, nw, :vr, bus_id)
     vi = _PMD.var(pm, nw, :vi, bus_id)
     crg = _PMD.var(pm, nw, :crg, id)
@@ -133,7 +133,7 @@ end
 
 "Simplified version of constraint_mc_current_balance_se, to perform state
 estimation with the reduced IVR formulation: no shunts, no storage elements, no active switches"
-function constraint_mc_current_balance_se(pm::_PMDSE.ReducedIVRPowerModel, i::Int; nw::Int=pm.cnw)
+function constraint_mc_current_balance_se(pm::_PMDSE.ReducedIVRUPowerModel, i::Int; nw::Int=_IM.nw_id_default)
 
     bus = _PMD.ref(pm, nw, :bus, i)
     bus_arcs = _PMD.ref(pm, nw, :bus_arcs_conns_branch, i)
@@ -141,14 +141,14 @@ function constraint_mc_current_balance_se(pm::_PMDSE.ReducedIVRPowerModel, i::In
     bus_gens = _PMD.ref(pm, nw, :bus_conns_gen, i)
     bus_loads = _PMD.ref(pm, nw, :bus_conns_load, i)
 
-    cr    = get(_PMD.var(pm, nw),    :cr, Dict()); _PMs._check_var_keys(cr, bus_arcs, "real current", "branch")
-    ci    = get(_PMD.var(pm, nw),    :ci, Dict()); _PMs._check_var_keys(ci, bus_arcs, "imaginary current", "branch")
-    crd   = get(_PMD.var(pm, nw),   :crd, Dict()); _PMs._check_var_keys(crd, bus_loads, "real current", "load")
-    cid   = get(_PMD.var(pm, nw),   :cid, Dict()); _PMs._check_var_keys(cid, bus_loads, "imaginary current", "load")
-    crg   = get(_PMD.var(pm, nw),   :crg, Dict()); _PMs._check_var_keys(crg, bus_gens, "real current", "generator")
-    cig   = get(_PMD.var(pm, nw),   :cig, Dict()); _PMs._check_var_keys(cig, bus_gens, "imaginary current", "generator")
-    crt   = get(_PMD.var(pm, nw),   :crt, Dict()); _PMs._check_var_keys(crt, bus_arcs_trans, "real current", "transformer")
-    cit   = get(_PMD.var(pm, nw),   :cit, Dict()); _PMs._check_var_keys(cit, bus_arcs_trans, "imaginary current", "transformer")
+    cr    = get(_PMD.var(pm, nw),    :cr, Dict()); _PMD._check_var_keys(cr, bus_arcs, "real current", "branch")
+    ci    = get(_PMD.var(pm, nw),    :ci, Dict()); _PMD._check_var_keys(ci, bus_arcs, "imaginary current", "branch")
+    crd   = get(_PMD.var(pm, nw),   :crd, Dict()); _PMD._check_var_keys(crd, bus_loads, "real current", "load")
+    cid   = get(_PMD.var(pm, nw),   :cid, Dict()); _PMD._check_var_keys(cid, bus_loads, "imaginary current", "load")
+    crg   = get(_PMD.var(pm, nw),   :crg, Dict()); _PMD._check_var_keys(crg, bus_gens, "real current", "generator")
+    cig   = get(_PMD.var(pm, nw),   :cig, Dict()); _PMD._check_var_keys(cig, bus_gens, "imaginary current", "generator")
+    crt   = get(_PMD.var(pm, nw),   :crt, Dict()); _PMD._check_var_keys(crt, bus_arcs_trans, "real current", "transformer")
+    cit   = get(_PMD.var(pm, nw),   :cit, Dict()); _PMD._check_var_keys(cit, bus_arcs_trans, "imaginary current", "transformer")
 
     terminals = bus["terminals"]
     grounded =  bus["grounded"]
@@ -173,7 +173,7 @@ end
 
 "Simplified version of constraint_mc_current_balance, to perform power flow
 calculations with the reduced IVR formulation: no shunts, no storage elements, no active switches"
-function constraint_mc_current_balance(pm::_PMDSE.ReducedIVRPowerModel, i::Int; nw::Int=pm.cnw)
+function constraint_mc_current_balance(pm::_PMDSE.ReducedIVRUPowerModel, i::Int; nw::Int=_IM.nw_id_default)
 
     bus = _PMD.ref(pm, nw, :bus, i)
     bus_arcs = _PMD.ref(pm, nw, :bus_arcs_conns_branch, i)
@@ -181,14 +181,14 @@ function constraint_mc_current_balance(pm::_PMDSE.ReducedIVRPowerModel, i::Int; 
     bus_gens = _PMD.ref(pm, nw, :bus_conns_gen, i)
     bus_loads = _PMD.ref(pm, nw, :bus_conns_load, i)
 
-    cr    = get(_PMD.var(pm, nw),    :cr, Dict()); _PMs._check_var_keys(cr, bus_arcs, "real current", "branch")
-    ci    = get(_PMD.var(pm, nw),    :ci, Dict()); _PMs._check_var_keys(ci, bus_arcs, "imaginary current", "branch")
-    crd   = get(_PMD.var(pm, nw),   :crd, Dict()); _PMs._check_var_keys(crd, bus_loads, "real current", "load")
-    cid   = get(_PMD.var(pm, nw),   :cid, Dict()); _PMs._check_var_keys(cid, bus_loads, "imaginary current", "load")
-    crg   = get(_PMD.var(pm, nw),   :crg, Dict()); _PMs._check_var_keys(crg, bus_gens, "real current", "generator")
-    cig   = get(_PMD.var(pm, nw),   :cig, Dict()); _PMs._check_var_keys(cig, bus_gens, "imaginary current", "generator")
-    crt   = get(_PMD.var(pm, nw),   :crt, Dict()); _PMs._check_var_keys(crt, bus_arcs_trans, "real current", "transformer")
-    cit   = get(_PMD.var(pm, nw),   :cit, Dict()); _PMs._check_var_keys(cit, bus_arcs_trans, "imaginary current", "transformer")
+    cr    = get(_PMD.var(pm, nw),    :cr, Dict()); _PMD._check_var_keys(cr, bus_arcs, "real current", "branch")
+    ci    = get(_PMD.var(pm, nw),    :ci, Dict()); _PMD._check_var_keys(ci, bus_arcs, "imaginary current", "branch")
+    crd   = get(_PMD.var(pm, nw),   :crd, Dict()); _PMD._check_var_keys(crd, bus_loads, "real current", "load")
+    cid   = get(_PMD.var(pm, nw),   :cid, Dict()); _PMD._check_var_keys(cid, bus_loads, "imaginary current", "load")
+    crg   = get(_PMD.var(pm, nw),   :crg, Dict()); _PMD._check_var_keys(crg, bus_gens, "real current", "generator")
+    cig   = get(_PMD.var(pm, nw),   :cig, Dict()); _PMD._check_var_keys(cig, bus_gens, "imaginary current", "generator")
+    crt   = get(_PMD.var(pm, nw),   :crt, Dict()); _PMD._check_var_keys(crt, bus_arcs_trans, "real current", "transformer")
+    cit   = get(_PMD.var(pm, nw),   :cit, Dict()); _PMD._check_var_keys(cit, bus_arcs_trans, "imaginary current", "transformer")
 
     terminals = bus["terminals"]
     grounded =  bus["grounded"]
@@ -211,12 +211,12 @@ function constraint_mc_current_balance(pm::_PMDSE.ReducedIVRPowerModel, i::Int; 
     end
 end
 
-"Constraint that defines the voltage drop along a branch for the ReducedIVRPowerModel.
+"Constraint that defines the voltage drop along a branch for the ReducedIVRUPowerModel.
 Conceptually similar to the same function in PowerModelsDistribution, but written in
 function of the total current instead of the series current.
-This is because in the ReducedIVRPowerModel series current is not defined as it is identical to
+This is because in the ReducedIVRUPowerModel series current is not defined as it is identical to
 the total current due to the absence of shunt admittance."
-function constraint_mc_bus_voltage_drop(pm::ReducedIVRPowerModel, i::Int; nw::Int=pm.cnw)
+function constraint_mc_bus_voltage_drop(pm::ReducedIVRUPowerModel, i::Int; nw::Int=_IM.nw_id_default)
 
     branch = _PMD.ref(pm, nw, :branch, i)
     f_bus = branch["f_bus"]
@@ -240,8 +240,8 @@ function constraint_mc_bus_voltage_drop(pm::ReducedIVRPowerModel, i::Int; nw::In
 end
 
 "This constraint makes sure that the current entering and exiting a branch are equivalent.
-It is used for state estimation and power flow calculations with the ReducedIVRPowerModel."
-function constraint_current_to_from(pm::ReducedIVRPowerModel, i::Int; nw::Int=pm.cnw)
+It is used for state estimation and power flow calculations with the ReducedIVRUPowerModel."
+function constraint_current_to_from(pm::ReducedIVRUPowerModel, i::Int; nw::Int=_IM.nw_id_default)
 
     branch = _PMD.ref(pm, nw, :branch, i)
     f_bus = branch["f_bus"]