+++ src/io/distributions.jl
@@ -2,7 +2,7 @@
 #  Copyright 2020, Marta Vanin, Tom Van Acker                                  #
 ################################################################################
 # PowerModelsDistributionStateEstimation.jl                                    #
-# An extention package of PowerModels(Distribution).jl for Static Power System #
+# An extention package of PowerModelsDistribution.jl for Static Power System   #
 # State Estimation.                                                            #
 ################################################################################
 
@@ -88,7 +88,7 @@ function mode(dst::ExtendedBeta)
     elseif α == 1 && β > 1  return min
     elseif α ≥ 1  && β < 1  return max
     elseif α > 1  && β == 1 return max
-    elseif α ≤ 1  && β ≤ 1  Memento.error(_LOGGER, "mode is defined only when α > 1 and/or β > 1")
+    elseif α ≤ 1  && β ≤ 1  error("mode is defined only when α > 1 and/or β > 1")
     end
 end
 skewness(dst::ExtendedBeta) =