+++ test/runtests.jl
@@ -9,16 +9,14 @@
 # using pkgs
 using Distributions, GaussianMixtures
 using Ipopt
-using PowerModels
-using PowerModelsDistribution
+using PowerModels, PowerModelsDistribution
 using PowerModelsDistributionStateEstimation
-#using SCS #removed while SDP tests are not e
+#using SCS #removed while SDP tests are not active
 using Test
 
 # pkg const
 const _DST = Distributions
 const _GMM = GaussianMixtures
-const _PMs = PowerModels
 const _PMD = PowerModelsDistribution
 const _PMDSE = PowerModelsDistributionStateEstimation
 
@@ -49,6 +47,7 @@ ipopt_solver = optimizer_with_attributes(Ipopt.Optimizer,"max_cpu_time"=>300.0,
     include("non_exact_forms.jl")
     include("power_flow.jl")
     include("pseudo_measurements.jl")
+    include("single_conductor_branches.jl")
     include("utils_and_start_val.jl")
     include("with_errors.jl")